fixed env
