+++ docker/server/entrypoint.sh
@@ -35,7 +35,6 @@ fi
 # Replace env var sentinel in dashboard build output
 declare -A replacements=(
   ["STACK_ENV_VAR_SENTINEL_USE_INLINE_ENV_VARS"]="true"
-  ["STACK_ENV_VAR_SENTINEL_NEXT_PUBLIC_STACK_PROJECT_ID"]="${NEXT_PUBLIC_STACK_PROJECT_ID}"
   ["STACK_ENV_VAR_SENTINEL_NEXT_PUBLIC_STACK_API_URL"]="${NEXT_PUBLIC_STACK_API_URL}"
   ["STACK_ENV_VAR_SENTINEL_NEXT_PUBLIC_CLIENT_STACK_API_URL"]="${NEXT_PUBLIC_CLIENT_STACK_API_URL}"
   ["STACK_ENV_VAR_SENTINEL_NEXT_PUBLIC_SERVER_STACK_API_URL"]="${NEXT_PUBLIC_SERVER_STACK_API_URL}"
@@ -48,6 +47,9 @@ declare -A replacements=(
   ["STACK_ENV_VAR_SENTINEL_NEXT_PUBLIC_VERSION_ALERTER_SEVERE_ONLY"]="${NEXT_PUBLIC_VERSION_ALERTER_SEVERE_ONLY}"
   ["STACK_ENV_VAR_SENTINEL_NEXT_PUBLIC_STACK_EMULATOR_ENABLED"]="${NEXT_PUBLIC_STACK_EMULATOR_ENABLED}"
   ["STACK_ENV_VAR_SENTINEL_NEXT_PUBLIC_STACK_EMULATOR_PROJECT_ID"]="${NEXT_PUBLIC_STACK_EMULATOR_PROJECT_ID}"
+  ["STACK_ENV_VAR_SENTINEL_NEXT_PUBLIC_STACK_PROJECT_ID"]="${NEXT_PUBLIC_STACK_PROJECT_ID}"
+  ["STACK_ENV_VAR_SENTINEL_NEXT_PUBLIC_STACK_PUBLISHABLE_CLIENT_KEY"]="${NEXT_PUBLIC_STACK_PUBLISHABLE_CLIENT_KEY}"
+  ["STACK_ENV_VAR_SENTINEL_NEXT_PUBLIC_STACK_URL"]="${NEXT_PUBLIC_STACK_URL}"
 )
 
 # Check for any unhandled sentinel values
